Add an option per-field-per-entity transform context.
The idea is to conveniently store "meta"-information about the
transformation.

Use case:

* multi-user encryption, where the user-list is stored alongside the
data in the same field
* when changing a value the list of allowed readers should stay the same
* and it should optionally be possible to change the list of readers

In this scenario the transformation context field would hold the list of
readers. When updating a value, the transformer implementation would
fetch the public keys of all users and do something like PHP's
openssl_seal() (but with a sensible encryption algorithm).